==> src/5.fuzzy.ipynb <==
### Imports

import os
import numpy as np
import pandas as pd
import skfuzzy as fuzz
from skfuzzy import control as ctrl

### Inputs

radnet_input_df = pd.read_csv("../inputs/radnet_input.csv")
radnet_input_efsr_edge = radnet_input_df['efsr_edge'].values

inference_output_df = pd.read_csv("../outputs/inference_output.csv")
inference_output_class = inference_output_df['class'].values
inference_output_confidence = inference_output_df['confidence'].values

edges_input_df = pd.read_csv("../inputs/edges_input.csv")
edges_input = edges_input_df[["rssi", "etx", "delay", "busy_fraction"]].values

### Fuzzy Logic

conf_model = ctrl.Antecedent(np.arange(0, 1.01, 0.01), 'conf_model')
conf_efsr = ctrl.Antecedent(np.arange(0, 1.01, 0.01), 'conf_efsr')
quality = ctrl.Consequent(np.arange(0, 1.01, 0.01), 'quality')

conf_model['baixa'] = fuzz.trimf(conf_model.universe, [0, 0, 0.5])
conf_model['media'] = fuzz.trimf(conf_model.universe, [0, 0.5, 1])
conf_model['alta'] = fuzz.trimf(conf_model.universe, [0.5, 1, 1])

conf_efsr['baixa'] = fuzz.trimf(conf_efsr.universe, [0, 0, 0.5])
conf_efsr['media'] = fuzz.trimf(conf_efsr.universe, [0, 0.5, 1])
conf_efsr['alta'] = fuzz.trimf(conf_efsr.universe, [0.5, 1, 1])

quality['ruim'] = fuzz.trimf(quality.universe, [0, 0, 0.5])
quality['media'] = fuzz.trimf(quality.universe, [0, 0.5, 1])
quality['boa'] = fuzz.trimf(quality.universe, [0.5, 1, 1])

rules = [
    ctrl.Rule(conf_model['alta'] & conf_efsr['alta'], quality['boa']),
    ctrl.Rule(conf_model['media'] & conf_efsr['alta'], quality['boa']),
    ctrl.Rule(conf_model['baixa'] & conf_efsr['alta'], quality['media']),
    ctrl.Rule(conf_model['alta'] & conf_efsr['media'], quality['boa']),
    ctrl.Rule(conf_model['media'] & conf_efsr['media'], quality['media']),
    ctrl.Rule(conf_model['baixa'] & conf_efsr['media'], quality['ruim']),
    ctrl.Rule(conf_model['alta'] & conf_efsr['baixa'], quality['media']),
    ctrl.Rule(conf_model['media'] & conf_efsr['baixa'], quality['ruim']),
    ctrl.Rule(conf_model['baixa'] & conf_efsr['baixa'], quality['ruim'])
]

fuzzy_system = ctrl.ControlSystemSimulation(ctrl.ControlSystem(rules))

confidence_efsr = np.clip((radnet_input_efsr_edge - 0.20) / (0.98 - 0.20), 0, 1) # Normalização
confidence_model = inference_output_confidence

fuzzy = np.zeros(len(confidence_model))

for i in range(len(confidence_model)):
    fuzzy_system.input['conf_model'] = confidence_model[i]
    fuzzy_system.input['conf_efsr'] = confidence_efsr[i]
    fuzzy_system.compute()
    fuzzy[i] = fuzzy_system.output['quality']

### Inference

cls = inference_output_class
conf = fuzzy

print(f"{'RSSI':<6} {'ETX':<5} {'Delay':<6} {'Busy':<6} {'EFSR':<6} {'Classe':<8} {'Status':<8} {'Confiança':<10}")
print(f"{'-'*62}")

CLASS_NAMES = {0: "Ruim", 1: "Bom"}

for i, row in enumerate(edges_input):
    print(f"{row[0]:<6.0f} {row[1]:<5.1f} {row[2]:<6.0f} {row[3]:<6.2f} "
          f"{radnet_input_efsr_edge[i]:<6.2f} "
          f"{cls[i]:<8} {CLASS_NAMES[int(cls[i])]:<8} {conf[i]:<10.2f}")

### Save

fuzzy_output = pd.DataFrame({
    'class': cls,
    'confidence': conf
})

os.makedirs("../outputs", exist_ok=True)
fuzzy_output.to_csv("../outputs/fuzzy_output.csv", index=False)

print("Lógica Fuzzy salva em ../outputs/fuzzy_output.csv")
