==> src/6.kalman.ipynb <==
### Imports

import os
import numpy as np
import pandas as pd
from filterpy.kalman import KalmanFilter

### Inputs

fuzzy_output_df = pd.read_csv("../outputs/fuzzy_output.csv")
fuzzy_output_class = fuzzy_output_df['class'].values
fuzzy_output_confidence = fuzzy_output_df['confidence'].values

edges_input_df = pd.read_csv("../inputs/edges_input.csv")
edges_input = edges_input_df[["rssi", "etx", "delay", "busy_fraction"]].values

radnet_input_df = pd.read_csv("../inputs/radnet_input.csv")
radnet_input_efsr_edge = radnet_input_df['efsr_edge'].values

### Kalman Filter

kf = KalmanFilter(dim_x=1, dim_z=1)

confidence_fusion = fuzzy_output_confidence

kf.x = np.array([confidence_fusion[0]])
kf.F = np.array([[1.]])
kf.H = np.array([[1.]])
kf.P = 1.0   # covariância inicial do estado
kf.R = 2e-2  # covariância da observação (measurement noise)
kf.Q = 1e-2  # covariância do processo (process noise)

kalman = []

for z in confidence_fusion:
    kf.predict()           
    kf.update(z)           
    kalman.append(kf.x[0])

kalman = np.array(kalman)

### Inference

cls = fuzzy_output_class
conf = kalman

print(f"{'RSSI':<6} {'ETX':<5} {'Delay':<6} {'Busy':<6} {'EFSR':<6} {'Classe':<8} {'Status':<8} {'Confiança':<10}")
print(f"{'-'*62}")

CLASS_NAMES = {0: "Ruim", 1: "Bom"}

for i, row in enumerate(edges_input):
    print(f"{row[0]:<6.0f} {row[1]:<5.1f} {row[2]:<6.0f} {row[3]:<6.2f} "
          f"{radnet_input_efsr_edge[i]:<6.2f} "
          f"{cls[i]:<8} {CLASS_NAMES[int(cls[i])]:<8} {conf[i]:<10.2f}")

### Save

kalman_output = pd.DataFrame({
    'class': cls,
    'confidence': conf
})

os.makedirs("../outputs", exist_ok=True)
kalman_output.to_csv("../outputs/kalman_output.csv", index=False)

print("Filtro de Kalman salvo em ../outputs/kalman_output.csv")
